Resolved CID 351632 (#1 of 1): Invalid type in argument to printf format specifier
invalid_type: Argument this->m_disabled_spells.size() to format specifier %u was expected to have type unsigned int but has type unsigned long

+++ src/world/Objects/ObjectMgr.cpp
@@ -2822,7 +2822,7 @@ void ObjectMgr::LoadDisabledSpells()
         delete result;
     }
 
-    LogNotice("ObjectMgr : %u disabled spells.", m_disabled_spells.size());
+    LogNotice("ObjectMgr : %u disabled spells.", static_cast<uint32_t>(m_disabled_spells.size()));
 }
 
 void ObjectMgr::ReloadDisabledSpells()
